automagic/requirements.txt
django==1.11.1
#MySQL-Python==1.2.3
django-users2==0.2.1
django-cors-headers==2.0.2
djangorestframework==3.6.3
django-filter==1.1.0
python-jenkins==0.4.14
selenium
requests
paramiko==2.1.2
scapy==2.3.3
demjson
django-filter==1.1.0
crontab
PyYAML
requests_toolbelt
HttpRunner==1.5.15
